requirements.txt
asgiref==3.5.2
Django==4.1.3
django-active-link==0.1.8
django-i18n==1.0.4
django-modeltranslation==0.18.5
django-simple-history==3.1.1
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2==2.9.5
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
